character-profile-matching/requirements.txt
asposestorage==1.0.2
matplotlib==3.2.2
numpy==1.21.2
Pillow==9.2.0
python_Levenshtein==0.12.2
torch==1.7.0
torchvision==0.8.1
lmdb==1.3.0